Make window maximized for "Run" page to allow onboard keyboard to appear
This is a Workaround for a bug in onboard on-screen keyboard not coming to front
in full screen mode. onboard stays behind a full screen window.
So, "Run" page of launcher would lack possibility to enter command on a touch
screen, which would make the page futile.
Reducing the launcher window to maximzed state lets the keyboard appear.
However, the taskbar gets visible.