Add focus and keyboard handling to "Run" page
This simplifies the handling with a virtual keyboard, especially, with the
autoshow feature of the onboard keyboard.

Note:
The (ApplicationModal) MessageBox, that appears when pressing the run button
and the command line is invalid, rarely produces a warning message on stderr,
when closed. This is of the form:
    qt.qpa.xcb: QXcbConnection: XCB error: 3 (BadWindow), sequence: 888,
    resource id: 16810974, major code: 40 (TranslateCoords), minor code: 0
where the numbers for sequence and resource id change and the rest repeats.

This a known bug in Qt, but does not crash the program. So it can be ignored.
See here: https://github.com/vgc/vgc/issues/6